requirements_dev.txt
faker==4.0.2
pylint==2.5.0
pylint-django==2.0.15
coverage==5.1
mypy==0.770
django-stubs==1.5.0
djangorestframework-stubs==1.2.0
